+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v4.1.0
 
-* `added` New method `client.getOrder()` to refresh order from CA
+* `added` Option `preferredChain` added to `client.getCertificate()` and `client.auto()` to indicate which certificate chain is preferred if a CA offers multiple
+    * Related: [https://community.letsencrypt.org/t/transition-to-isrgs-root-delayed-until-jan-11-2021/125516](https://community.letsencrypt.org/t/transition-to-isrgs-root-delayed-until-jan-11-2021/125516)
+* `added` Method `client.getOrder()` to refresh order from CA
 * `fixed` Upgrade `axios@0.21.0`
 * `fixed` Error when attempting to revoke a certificate chain
 * `fixed` Missing URL augmentation in `client.finalizeOrder()` and `client.deactivateAuthorization()`